requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==25.4.0
boto3==1.42.91
botocore==1.42.91
certifi==2026.4.22
cffi==2.0.0
Django==6.0.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
minio==7.2.18
neuon==0.0.4.4
pillow==12.2.0
psycopg2==2.9.11
pycparser==3.0
pycryptodome==3.23.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
style==1.1.0
typing_extensions==4.15.0
update==0.0.1
uritemplate==4.2.0
urllib3==2.6.3
requests==2.33.1
